refactor: move per namespace reports aggregator in a sub package
