+++ pkg/controllers/report/aggregate/namespace/log.go
@@ -1,4 +1,4 @@
-package aggregate
+package namespace
 
 import "github.com/kyverno/kyverno/pkg/logging"
